move most definitions to kernel crate & update bench CLI
reorg is needed at some point
