+++ benches/src/main.rs
@@ -9,9 +9,12 @@ use tikv_jemallocator::Jemalloc;
 use honeycomb_benches::{
     cli::{Benches, Cli, Format},
     cut_edges::bench_cut_edges,
+    // delaunay::bench_delaunay,
     grid_gen::bench_generate_2d_grid,
     grisubal::bench_grisubal,
-    prof_init, prof_start, prof_stop,
+    prof_init,
+    prof_start,
+    prof_stop,
     remesh::bench_remesh,
     shift::bench_shift,
 };
@@ -72,6 +75,7 @@ fn run_benchmarks<T: CoordsFloat>(cli: Cli) {
     let map: CMap2<T> = match cli.benches {
         Benches::Generate2dGrid(args) => bench_generate_2d_grid(args),
         Benches::CutEdges(args) => bench_cut_edges(args),
+        Benches::DelaunayBox(_args) => todo!(), // bench_delaunay(args),
         Benches::Grisubal(args) => bench_grisubal(args),
         Benches::Remesh(args) => bench_remesh(args),
         Benches::Shift(args) => bench_shift(args),
